rework tabs

+++ src/shared/api/entityApi/Essence.ts
@@ -58,7 +58,7 @@ export default class Essense<T extends { id?: IdentityType }> {
     try {
       this.updateEntity(id, data);
 
-      if (!entity.isFake) {
+      if (!entity._isFake) {
         return await this.api.patch(`${this.path}/${id}`, restData);
       }
     } catch (error) {
@@ -85,15 +85,19 @@ export default class Essense<T extends { id?: IdentityType }> {
   }
 
   async delete(id: IdentityType, { deleteFlag }: DeleteOptions = {}) {
-    
+    const entity = this.selector.getById(this.store.getState(), id);
+
     if (deleteFlag) {
       this.store.dispatch(updateEntity({ id, type: this.entityName, data: { _isDeleted: true } }));
     } else {
       this.store.dispatch(deleteEntity({ id, type: this.entityName }));
     }
-    const result = await this.api.delete<void>(`${this.path}/${id}`);
 
-    return result;
+    if (entity._isFake) {
+      return;
+    }
+
+    await this.api.delete<void>(`${this.path}/${id}`);
   }
 
   updateEntity(id: IdentityType, data: Partial<T>) {
